# model/metadata: make all fields non-pointer types