… beater/api/profile/handler_test.go
… model/error/event_test.go
… model/metadata/kubernetes_test.go
… model/metadata/metadata_test.go
… model/metadata/process_test.go
… model/metadata/service_test.go
… model/metadata/user_test.go
… model/metricset/event_test.go
… model/modeldecoder/context_test.go
… model/modeldecoder/error_test.go
… model/modeldecoder/kubernetes_test.go
… model/modeldecoder/metadata_test.go
… model/modeldecoder/metricset_test.go
… model/modeldecoder/process.go
… model/modeldecoder/process_test.go
… model/modeldecoder/service_test.go
… model/modeldecoder/span.go
… model/modeldecoder/span_test.go
… model/modeldecoder/system.go
… model/modeldecoder/system_test.go
… model/modeldecoder/test_approved_model/context_full_event with experimental=true.approved.json
… model/modeldecoder/test_approved_model/context_input_nil.approved.json
… model/modeldecoder/test_approved_model/context_user_id_integer.approved.json
… model/modeldecoder/transaction.go
… model/modeldecoder/transaction_test.go
… model/modeldecoder/user_test.go
… model/modeldecoder/utility.go
… model/profile/profile_test.go
… model/span/event.go
… model/span/event_test.go